separate BluepySimulationConfig and SonataSimulationConfig into different modules
