+++ bluecellulab/circuit/config/__init__.py
@@ -1,3 +1,5 @@
 """Simulation configuration file parser module."""
 
-from .simulation_config import BluepySimulationConfig, SimulationConfig, SonataSimulationConfig
+from .definition import SimulationConfig
+from .bluepy_simulation_config import BluepySimulationConfig
+from .sonata_simulation_config import SonataSimulationConfig
